requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.7

# WebSocket and Async Support
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3

# HTTP Requests and Utilities
requests==2.31.0
urllib3==1.26.18
certifi>=2023.7.22

# WSGI Server for Production
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# Twilio SDK
twilio==8.10.0

# OpenAI SDK
openai==1.3.0

# Additional Production Dependencies
wheel>=0.38.0
setuptools>=65.0.0

# Security and Performance
cryptography>=41.0.0
